+++ lib/SelfBalancingRobot/src/SBR_Telemetry.cpp
@@ -38,10 +38,10 @@ int packTelemetryData_MPC(uint8_t* telemetryDataPtr, uint32_t id, const MotorPai
     td->id = id;
     td->type = TD_MPC::TYPE;
     td->len = sizeof(TD_MPC);
-    td->tickInterval = motorPairController.getTickCountDelta();
+    td->tickInterval = static_cast<uint8_t>(motorPairController.getTickCountDelta());
 
     td->flags = motorPairController.motorsIsOn() ? TD_MPC::MOTORS_ON_FLAG : 0x00;
-    td->flags |= (TD_MPC::CONTROL_MODE_MASK & motorPairController.getControlMode());
+    td->flags |= static_cast<uint8_t>(TD_MPC::CONTROL_MODE_MASK & motorPairController.getControlMode());
 
     motor_pair_controller_telemetry_t telemetryData;
     motorPairController.getTelemetryData(telemetryData);
